Utilize Aggregator and CounterCache
I'm running an experiment now, and it seems this change doesn't
actually change the metric names visibly, and the change is more
behind the scenes, e.g.:

This is a relatively obvious change, and there's little to no
downside just doing this.

```ruby
10.times do |i|
  Pliny::Metrics.count "foo", i
end
```

Results in a single submission within a tick of:

```
{name: "foo", value: 45}
```

```ruby
10.times do |i|
  Pliny::Metrics.measure "foo" do
    sleep i
  end
end
```

That would submit a single measurement as follows:

```
{name: "foo", count: 10, sum: 45, min: 0, max: 9}
```

And that would all be accessible from your metric page by
choosing the appropriate summary function.

Contentious points:

- Should we just copy `CounterCache` from librato-rack, or do
  we add it as a dependency. I do not have a strong opinion
  on this one, having done more go recently, the "copy what
  you only need" approach is more frequently favored.
- Should we try to still use Queue as before? I'm leaning heavily
  to just stop doing it, as we are just unnecessarily bumping
  our rate limit use without much benefit (e.g. we'd still be using
  SSA regardless, the net numbers would be the same, etc)